Refactoring DecalEmitters
